requirements.txt
aiohttp==3.7.4.post0
aioredis==1.3.1
amqp==5.0.6
aniso8601==7.0.0
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
backports.entry-points-selectable==1.1.0
billiard==3.6.4.0
celery==5.1.2
cffi==1.14.6
cfgv==3.3.1
channels==3.0.4
channels-redis==3.3.0
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==3.4.8
daphne==3.0.2
distlib==0.3.3
Django==3.2.7
django-channels-graphql-ws==0.8.0
django-cors-headers==3.8.0
django-filter==2.4.0
filelock==3.1.0
graphene==2.1.9
graphene-django==2.15.0
graphene-file-upload==1.3.0
graphql-core==2.3.2
graphql-relay==2.0.1
hiredis==2.0.0
hyperlink==21.0.0
identify==2.2.15
idna==3.2
incremental==21.3.0
kombu==5.1.0
msgpack==1.0.2
multidict==5.1.0
nodeenv==1.6.0
Pillow==8.3.2
platformdirs==2.4.0
pre-commit==2.15.0
promise==2.3
prompt-toolkit==3.0.20
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.1.0
pyOpenSSL==20.0.1
python-decouple==3.4
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
rsa==4.7.2
Rx==1.6.1
service-identity==21.1.0
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
Twisted==21.7.0
txaio==21.2.1
typing-extensions==3.10.0.2
vine==5.0.0
virtualenv==20.8.1
wcwidth==0.2.5
yarl==1.6.3
zope.interface==5.4.0
